fix: align risk restoration with review feedback
